+++ app/Console/Kernel.php
@@ -17,6 +17,7 @@
 
 namespace App\Console;
 
+use App\Console\Commands\ProBINDImportZone;
 use App\Console\Commands\ProBINDPushZones;
 use Illuminate\Console\Scheduling\Schedule;
 use Illuminate\Foundation\Console\Kernel as ConsoleKernel;
@@ -31,6 +32,7 @@ class Kernel extends ConsoleKernel
      */
     protected $commands = [
         ProBINDPushZones::class,
+        ProBINDImportZone::class
     ];
 
     /**