+++ app/Http/Controllers/ZoneController.php
@@ -32,10 +32,7 @@ class ZoneController extends Controller
      */
     public function index()
     {
-        $zones = Zone::all();
-
-        return view('zone.index')
-            ->with('zones', $zones);
+        return view('zone.index');
     }
 
     /**
@@ -58,12 +55,17 @@ class ZoneController extends Controller
     public function store(ZoneCreateRequest $request)
     {
         $zone = new Zone();
+        $zone->domain = $request->input('domain');
+        $zone->reverse_zone = Zone::validateReverseDomainName($zone->domain);
 
-        // assign new serial and flag as updated
-        if ($request->input('type') == 'master') {
-            $zone->serial = Zone::createSerialNumber();
-            $zone->updated = true;
+        // if it's a Master zone, assign new Serial Number and flag pending changes.
+        if (!$request->has('master_server')) {
+            $zone->serial = Zone::generateSerialNumber();
+            $zone->has_modifications = true;
         }
+
+        // deal with checkboxes
+        $zone->custom_settings = $request->has('custom_settings');
 
         $zone->fill($request->all())->save();
 
@@ -101,17 +103,18 @@ class ZoneController extends Controller
      * Update the specified resource in storage.
      *
      * @param  ZoneUpdateRequest $request
-     * @param  Zone $zone
+     * @param  Zone              $zone
      *
      * @return \Illuminate\Http\RedirectResponse
      */
     public function update(ZoneUpdateRequest $request, Zone $zone)
     {
+        // if it's a Master zone, assign new Serial Number and flag pending changes.
         if ($zone->isMasterZone()) {
-            // assign new serial and flag as updated
-            $zone->setSerialNumber();
+            $zone->getNewSerialNumber();
             $zone->setPendingChanges(true);
         }
+
         // deal with checkboxes
         $zone->custom_settings = $request->has('custom_settings');
 
@@ -161,15 +164,15 @@ class ZoneController extends Controller
         $zones = Zone::select([
             'id',
             'domain',
-            'master',
-            'updated'
+            'master_server',
+            'has_modifications'
         ]);
 
         return $dataTable::of($zones)
             ->addColumn('type', function (Zone $zone) {
                 return trans('zone/model.types.' . $zone->getTypeOfZone());
             })
-            ->editColumn('updated', function (Zone $zone) {
+            ->editColumn('has_modifications', function (Zone $zone) {
                 return trans_choice('general.boolean', intval($zone->hasPendingChanges()));
             })
             ->addColumn('actions', function (Zone $zone) {