+++ app/Http/Requests/ZoneCreateRequest.php
@@ -39,7 +39,7 @@ class ZoneCreateRequest extends Request
     {
         return [
             'domain'          => 'required|string|unique:zones',
-            'master'          => 'sometimes|required|ip',
+            'master_server'   => 'sometimes|required|ip',
             'custom_settings' => 'sometimes|boolean',
             'refresh'         => 'required_if:custom_settings,1|integer|min:0|max:2147483647',
             'retry'           => 'required_if:custom_settings,1|integer|min:0|max:2147483647',